plotting completed
